nand2tetris: implement 3-gate Xor
The logical representation of Xor is (a AND not b) OR (not a AND b),
which, using de Morgan's law, reduces to (a OR b) AND (not a OR not b).
Since NAND is equivalent to NOT OR, the final expression is (a OR b) AND
(a NAND b).

This uses two fewer gates than the five gate initial solution.