nand2tetris: implement half-adder
Implemented the Half-Adder chip with an Xor and Add. The Xor controls
the sum output, and the And controls the carry output.

The truth table for the half-:
a   b | c   s
-------------
0   0 | 0   0
1   0 | 0   1
0   1 | 0   1
1   1 | 1   0

As you can see the carry column has the same truth table as the Add
gate and sum as the Xor.