nand2tetris: pass NestedCall test
The diff for this commit is large as it includes a refactor from a
String-based to an enum-based approach. This allows clearer expression
of desired commands. After refactoring and fixing bugs, the VM
translator now passes the NestedCall test.

In addition to passing the test, this commit adds functionality for
parsing an entire directory of VM files and omitting the Sys.init header
required for later tests.